LayoutTests/fast/block/float/hit-test-on-overlapping-floats.html

<!DOCTYPE html>
<html>
    <head>
        <style>
        #leftc {
            float: left;
            width: 500px;
            background-color: silver;
            margin-right: -100px;
        }

        #rightc {
            float: right;
            width: 400px;
            background-color: rgba(0, 128, 0, 0.75);
            padding-bottom: 1px;
        }
        </style>
    </head>
    <body>
        <div style="width:800px">
            <div id="leftc">foo</div>
            <div id="rightc">bar</div>
            <div style="height:0px; width:0px; clear:left;"></div>
        </div>
        <script src="../../../resources/js-test-pre.js"></script>
        <script>
        shouldBe("document.elementFromPoint(450, 10)", "document.getElementById('rightc')");
        </script>
        <script src="../../../resources/js-test-post.js"></script>
	</body>
</html>
