Safari-609.1.11.1/LayoutTests/fast/block/float/hit-test-on-overlapping-floats-expected.txt

PASS document.elementFromPoint(450, 10) is document.getElementById('rightc')
PASS successfullyParsed is true

TEST COMPLETE
foobar
